pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "my-agent"
version = "0.1.0"
dependencies = [
  "python-dotenv",
  "google-adk==1.16.0",
  "google-cloud-aiplatform[adk,agent-engines]==1.126.1",
  "google-cloud-discoveryengine",
  "litellm",
  "absl-py",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["my_agent"]
